+++ src/ui/about.rs
@@ -104,7 +104,7 @@ impl SimpleComponent for AboutDialog {
 
                 "<ul>",
                     "<li>List missing dependencies on non-standard distros during initial setup</li>",
-                    "<li>Added 2.1.0 voiceovers sizes</li>",
+                    "<li>Added 2.2.0 voiceovers sizes</li>",
                 "</ul>",
 
                 "<p>Fixed</p>",