+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 17.06.2024
+
 ### Added
 
 - List missing dependencies on non-standard distros during initial setup
@@ -293,7 +295,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <br>
 
-[unreleased]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.5.5...next
+[unreleased]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.6.0...next
+[1.6.0]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.5.5...1.6.0
 [1.5.5]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.5.4...1.5.5
 [1.5.4]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.5.3...1.5.4
 [1.5.3]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.5.2...1.5.3